Test-code clean up
