+++ tests/CMakeLists.txt
@@ -1,4 +1,7 @@
 add_subdirectory(benchmarking)
+add_subdirectory(performance)
 add_subdirectory(test-utils)
 add_subdirectory(unit-tests)
 add_subdirectory(visualization)
+
+find_package(Python3 COMPONENTS Interpreter REQUIRED)
