+++ tests/benchmarking/CMakeLists.txt
@@ -1,9 +1,8 @@
 add_executable(benchmarking
         TatumQuantizationFitBenchmarking.cpp
+        RocTest.cpp
 )
 
 target_link_libraries(benchmarking PRIVATE
         test-utils
 )
-
-find_package(Python3 COMPONENTS Interpreter REQUIRED)
