+++ tests/benchmarking/TatumQuantizationFitBenchmarking.cpp
@@ -4,7 +4,6 @@
 #include "TestLteAudioReader.h"
 
 #include <catch2/catch_test_macros.hpp>
-#include <chrono>
 #include <cmath>
 #include <filesystem>
 #include <fstream>
@@ -14,110 +13,6 @@
 
 namespace LTE
 {
-namespace
-{
-const auto datasetRoot =
-   std::string(CMAKE_SOURCE_DIR) + "/tests/benchmarking/dataset";
-
-std::vector<std::string> GetBenchmarkingAudioFiles()
-{
-   const auto pythonScriptPath =
-      std::string(CMAKE_SOURCE_DIR) +
-      "/tests/benchmarking/download-benchmarking-dataset.py";
-   const auto command = "python " + pythonScriptPath + " " + datasetRoot;
-   const auto returnCode = system(command.c_str());
-   if (returnCode != 0)
-      throw std::runtime_error("Failed to download benchmarking dataset!");
-
-   std::vector<std::string> files;
-   namespace fs = std::filesystem;
-   for (const auto& entry : fs::directory_iterator(datasetRoot))
-      for (const auto& subEntry : fs::recursive_directory_iterator(entry))
-         if (
-            subEntry.is_regular_file() && subEntry.path().extension() == ".mp3")
-            files.push_back(subEntry.path().string());
-   std::sort(files.begin(), files.end());
-   return files;
-}
-
-std::string Pretty(const std::string& filename)
-{
-   // Remove the dataset root from the filename ...
-   const auto datasetRootLength = datasetRoot.length();
-   auto tmp = filename.substr(datasetRootLength + 1);
-   // ... and now the .mp3 extension:
-   tmp = tmp.substr(0, tmp.length() - 4);
-   // Replace backslashes with forward slashes:
-   std::replace(tmp.begin(), tmp.end(), '\\', '/');
-   return tmp;
-}
-
-bool isMP3File(const std::string& filename)
-{
-   return filename.substr(filename.find_last_of(".") + 1) == "mp3";
-}
-
-} // namespace
-
-TEST_CASE("GetRocInfo")
-{
-   // We use the AUC as a measure of the classifier's performance. With a
-   // suitable data set, this helps us detect regressions, and guide fine-tuning
-   // of the algorithm. This test should help understand how it works and also
-   // make sure that we've implemented that metric correctly :)
-
-   struct Sample
-   {
-      bool truth;
-      double score;
-   };
-
-   using Samples = std::vector<Sample>;
-
-   struct Expected
-   {
-      double areaUnderCurve;
-      double threshold;
-   };
-
-   struct TestCase
-   {
-      const Samples samples;
-      const double allowedFalsePositiveRate;
-      const Expected expected;
-   };
-
-   const std::vector<TestCase> testCases {
-      // Classifier is upside down. We don't tolerate false positives. The
-      // returned threshold is then 100 will satisfy this, but the TPR will also
-      // be 0 ...
-      TestCase { Samples { { true, 100. }, { false, 200. } }, 0.,
-                 Expected { 0., 200. } },
-
-      // Classifier is still upside down. We'll get true positives only if we
-      // accept an FPR of 1.
-      TestCase { Samples { { true, 100. }, { false, 200. } }, 1.,
-                 Expected { 0., 100. } },
-
-      // Now we have a classifier that works. We don't accept false positives.
-      TestCase { Samples { { false, 100. }, { false, 150. }, { true, 200. } },
-                 0., Expected { 1., 175. } },
-
-      // A random classifier, which should have an AUC of 0.75.
-      TestCase {
-         Samples { { false, 1. }, { true, 2. }, { false, 3. }, { true, 4. } },
-         0.5, Expected { .75, 1.5 } },
-   };
-
-   for (const auto& testCase : testCases)
-   {
-      const auto roc =
-         GetRocInfo(testCase.samples, testCase.allowedFalsePositiveRate);
-      REQUIRE(roc.areaUnderCurve == testCase.expected.areaUnderCurve);
-      REQUIRE(roc.threshold == testCase.expected.threshold);
-   }
-}
-
 TEST_CASE("GetChecksum")
 {
    constexpr auto bufferSize = 5;
@@ -183,7 +78,6 @@ TEST_CASE("TatumQuantizationFitBenchmarking")
    std::vector<Sample> samples;
    const auto numFiles = audioFiles.size();
    auto count = 0;
-   std::chrono::milliseconds computationTime { 0 };
    std::transform(
       audioFiles.begin(), audioFiles.begin() + numFiles,
       std::back_inserter(samples),
@@ -194,12 +88,8 @@ TEST_CASE("TatumQuantizationFitBenchmarking")
          checksum += GetChecksum(audio);
          QuantizationFitDebugOutput debugOutput;
          std::function<void(double)> progressCb;
-         const auto now = std::chrono::steady_clock::now();
          GetBpm(audio, lenientTolerance, progressCb, &debugOutput);
 
-         computationTime +=
-            std::chrono::duration_cast<std::chrono::milliseconds>(
-               std::chrono::steady_clock::now() - now);
          ProgressBar(progressBarWidth, 100 * count++ / numFiles);
          const auto expected = GetBpmFromFilename(audioFile);
          const auto truth = expected.has_value();
@@ -220,11 +110,6 @@ TEST_CASE("TatumQuantizationFitBenchmarking")
                         << Pretty(audioFile) << "\n";
          return Sample { truth, debugOutput.score, error };
       });
-
-   {
-      std::ofstream timeMeasurementFile { "./timeMeasurement.txt" };
-      timeMeasurementFile << computationTime.count() << "ms\n";
-   }
 
    // AUC of ROC curve. Tells how good our loop/not-loop clasifier is.
    const auto rocInfo = GetRocInfo(
