+++ tests/performance/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_executable(performance
+        TatumQuantizationFitPerformance.cpp
+)
+
+target_link_libraries(performance PRIVATE
+        test-utils
+)
